fix(core): move logic out of lib.rs files into modules
